Create a dropdown menu

Create a dropdown menu

fdDropdownMenu(..., type = c("messages", "notifications", "tasks"),
  badgeStatus = "primary", icon = NULL, .list = NULL)

Arguments

...

Items to put in the menu. Typically, message menus should contain fdMessages, notification menus should contain fdNotifications.

type

The type of menu. Should be one of "messages", "notifications", "tasks".

badgeStatus

The status of the badge which displays the number of items in the menu. This determines the badge's color. A value of NULL means to not display a badge.

icon

An icon to display in the header. By default, the icon is automatically selected depending on type, but it can be overriden with this argument.

.list

An optional list containing items to put in the menu Same as the ... arguments, but in list format. This can be useful when working with programmatically generated items.

See also

fdHeader for example usage.