Create an info box.

An info box displays a large icon on the left side, and a title, value (usually a number), and an optional smaller subtitle on the right side. Info boxes are meant to be placed in the main body of a dashboard.

fdInfoBox(title, value = NULL, subtitle = NULL,
  icon = fdIcon("bar-chart"), color = "aqua", width = 4, href = NULL,
  fill = FALSE)

Arguments

title

Title text.

value

The value to display in the box. Usually a number or short text.

subtitle

Subtitle text (optional).

icon

An icon tag, created by icon.

color

A color for the box. Valid colors are listed in validColors.

width

The width of the box, using the Bootstrap grid system. This is used for row-based layouts. The overall width of a region is 12, so the default valueBox width of 4 occupies 1/3 of that width. For column-based layouts, use NULL for the width; the width is set by the column that contains the box.

href

An optional URL to link to.

fill

If FALSE (the default), use a white background for the content, and the color argument for the background of the icon. If TRUE, use the color argument for the background of the content; the icon will use the same color with a slightly darkened background.

See also

fdBox for usage examples.

Other boxes: fdBox, fdValueBox

Examples

library(flightdeck)
library(htmltools)

# Example 1
tagList(
  fdInfoBox('Adjusted R-squared', '86.9%',
    color = 'blue', icon = fdIcon('chart-line', lib = 'entypo')),
  fdInfoBox('F-Statistic', 13.93,  'on 10 and 21 degrees of freedom',
    color = 'green', icon = fdIcon('star', lib = 'entypo')),
  fdInfoBox('Residual Standard Error', 2.65, 'on 21 degrees of freedom',
    color = 'purple', icon = fdIcon('flash', lib = 'entypo'))
) %>%
  fdPreview(wrapBox = F)

# Example 2: Interactive help embedded in stat box.
# The message supports markdown.
msg <- "The adjusted R-squared compares the explanatory power of regression models that contain different numbers of predictors.

Suppose you compare a five-predictor model with a higher R-squared to a one-predictor model. Does the five predictor model have a higher R-squared because it’s better? Or is the R-squared higher because it has more predictors? Simply compare the adjusted R-squared values to find out!

The adjusted R-squared is a modified version of R-squared that has been adjusted for the number of predictors in the model. The adjusted R-squared increases only if the new term improves the model more than would be expected by chance. It decreases when a predictor improves the model by less than expected by chance. The adjusted R-squared can be negative, but it’s usually not.  It is always lower than the R-squared."

adjRsquared <- span("Adjusted R-squared",
  fdIcon('info-circled', lib = 'entypo') %>%
    fdModal(msg, title = 'Adjusted R-Squared')
)

tagList(
  fdInfoBox(adjRsquared, '86.9%',
    color = 'blue', icon = fdIcon('chart-line', lib = 'entypo')),
  fdInfoBox('F-Statistic', 13.93,  'on 10 and 21 degrees of freedom',
    color = 'green', icon = fdIcon('star', lib = 'entypo')),
  fdInfoBox('Residual Standard Error', 2.65, 'on 21 degrees of freedom',
    color = 'purple', icon = fdIcon('flash', lib = 'entypo'))
) %>%
  fdPreview(wrapBox = F)