Display performance metrics by class for a classification model

Display performance metrics by class for a classification model

fdPanelClassificationMetricsByClass(actual, predicted, vertical = TRUE,
  cutoffs = c(0, 0.6, 0.75, 1))

computeClassificationMetricsByClass(actual, predicted, vertical = TRUE)

Arguments

actual

vector of actual classes.

predicted

vector of predicted classes.

vertical

boolean indicating if panel should display measures in columns.

cutoffs

vector of cutoffs to use for deciding bar colors.

See also

Other classification: fdPanelClassificationMetrics

Examples

library(rpart)
fit <- rpart(Species ~ ., data = iris)
actual <- iris$Species
predicted <- predict(fit, type = 'class')

computeClassificationMetricsByClass(actual = actual, predicted = predicted)

#>     Measure setosa versicolor virginica microAverage macroAverage
#> 1  Accuracy     NA         NA        NA    0.9733333    0.9600000
#> 2  F1_Score      1  0.9423077 0.9375000    0.9600000    0.9599359
#> 3 Precision      1  0.9074074 0.9782609    0.9600000    0.9618894
#> 4    Recall      1  0.9800000 0.9000000    0.9600000    0.9600000

if (interactive()){
  fdPanelClassificationMetricsByClass(actual = actual, predicted = predicted) %>%
    fdPreview("Classification Metrics")
}